Fix KV environment parse error
Some U-boot environments are padded with zeros and some are padded with
0xff characters. U-Boot tools pads with zeros and fwup pads with 0xff.
The existing code only worked with fwup padding. This was due to the environment
scanning code looking for exactly two 0 bytes to terminate the list of
kv pairs. If U-Boot generated the environment, there'd be more than two
0 bytes and the parsing step would break.

This commit fixes the issue and adds additional unit tests to verify the
environment parser. Binary environment files are checked in for
convenience so that users don't need fwup and U-boot tools installed to
run the tests.